+++ grid/app/main/sfl/processes/config.py
@@ -16,7 +16,7 @@ class Config(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     config = db.Column(db.PickleType)
     is_server_config = db.Column(db.Boolean, default=False)
-    fl_process_id = db.Column(db.BigInteger, db.ForeignKey("static_fl_process_.id"))
+    fl_process_id = db.Column(db.Integer, db.ForeignKey("static_fl_process_.id"))
 
     def __str__(self):
         return f"<Config id: {self.id} , configs: {self.config}>"
